+++ src/bika/lims/content/analysisrequest.py
@@ -19,6 +19,7 @@
 # Some rights reserved, see README and LICENSE.
 
 import base64
+import copy
 import functools
 import re
 from datetime import datetime
@@ -1435,6 +1436,13 @@ schema = BikaSchema.copy() + Schema((
             render_own_label=True,
         ),
     ),
+
+    # Selected analytes from multi-component analyses on Sample registration
+    RecordsField(
+        "ServiceAnalytes",
+        widget=ComputedWidget(visible=False)
+    ),
+
 ))
 
 
@@ -2623,6 +2631,23 @@ class AnalysisRequest(BaseFolder, ClientAwareMixin):
             "sort_order": "ascending",
         }
         return query
+
+    def getServiceAnalytesFor(self, service_or_uid):
+        """Return a list of dicts representing the analytes selected for this
+        sample and the given service. These analytes are usually selected on
+        sample registration
+        """
+        analytes = []
+        true_values = ("true", "1", "on", "True", True, 1)
+        target_uid = api.get_uid(service_or_uid)
+        all_analytes = self.getServiceAnalytes() or []
+        all_analytes = copy.deepcopy(all_analytes)
+        for analyte in all_analytes:
+            if analyte.get("uid") != target_uid:
+                continue
+            if analyte.get("value", None) in true_values:
+                analytes.append(analyte)
+        return analytes
 
 
 registerType(AnalysisRequest, PROJECTNAME)